Add select test
The original lebench includes this test and it will be useful to
Symbiote, add it back.

Note that this test departs from the method used by the others for
recording run times. Instead of doing the file I/O in the test loop, it
collects the results in a large array and then writes them out after the
test iterations are finished.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>